driver/src/requirements.txt
black==22.8.0
flake8==5.0.4
flake8-return==1.1.3
isort==5.10.1
environs==9.5.0
mypy==0.971
pytest==7.1.3
click==8.1.3
coloredlogs==15.0.1
pulsar-client==2.10.1
